src/data_loader.py
"""
Data Loading for BEAM-Net
===========================
Loads event-based datasets using the tonic library, which provides
standardized access to neuromorphic vision datasets.

Supported datasets:
  - N-MNIST (Orchard et al., 2015): Neuromorphic MNIST, 34×34 DVS sensor
  - DVS Gesture (Amir et al., 2017): 11 hand gesture classes from DVS128
  - N-Caltech101 (Orchard et al., 2015): Neuromorphic Caltech-101

These are real DVS recordings, not synthetic conversions — critical for
validating BEAM-Net's native event processing claims (§3.1).

Data flow:
  Raw events → tonic transforms (temporal binning) → PyTorch DataLoader
"""

import torch
from torch.utils.data import DataLoader, random_split
import numpy as np
from typing import Tuple, Optional


def get_dataloaders(cfg: dict) -> Tuple[DataLoader, DataLoader, DataLoader]:
    """
    Create train/val/test dataloaders for the configured dataset.
    
    Uses tonic library for DVS dataset loading with temporal binning.
    Falls back to torchvision MNIST if tonic is unavailable.
    
    Parameters
    ----------
    cfg : dict
        Full experiment config.
    
    Returns
    -------
    train_loader, val_loader, test_loader : DataLoader
    """
    dataset_name = cfg["data"]["dataset"]
    data_dir = cfg["data"]["data_dir"]
    batch_size = cfg["data"]["batch_size"]
    n_workers = cfg["data"]["num_workers"]
    n_bins = cfg["data"]["n_time_bins"]
    
    try:
        import tonic
        import tonic.transforms as T
        
        # Temporal binning transform: events → (n_bins, 2, H, W) frames
        sensor_size = _get_sensor_size(dataset_name)
        transform = tonic.transforms.Compose([
            tonic.transforms.Denoise(filter_time=10000),  # Remove noise
            tonic.transforms.ToFrame(
                sensor_size=sensor_size,
                n_time_bins=n_bins,
            ),
            torch.tensor,  # Convert to tensor
        ])
        
        if dataset_name == "nmnist":
            train_ds = tonic.datasets.NMNIST(
                save_to=data_dir, train=True, transform=transform
            )
            test_ds = tonic.datasets.NMNIST(
                save_to=data_dir, train=False, transform=transform
            )
        elif dataset_name == "dvs_gesture":
            train_ds = tonic.datasets.DVSGesture(
                save_to=data_dir, train=True, transform=transform
            )
            test_ds = tonic.datasets.DVSGesture(
                save_to=data_dir, train=False, transform=transform
            )
        else:
            # Fallback to torchvision MNIST for non-DVS datasets
            train_ds, test_ds = _fallback_mnist(data_dir)
            collate = None
        
        # Custom collate for variable-length event data
        collate = _event_collate_fn
        
    except ImportError:
        print("[Data] tonic not available, falling back to MNIST flat images")
        train_ds, test_ds = _fallback_mnist(data_dir)
        collate = None
    
    # Train/val split
    val_size = int(len(train_ds) * cfg["data"]["val_fraction"])
    train_size = len(train_ds) - val_size
    train_ds, val_ds = random_split(
        train_ds, [train_size, val_size],
        generator=torch.Generator().manual_seed(cfg["experiment"]["seed"])
    )
    
    train_loader = DataLoader(
        train_ds, batch_size=batch_size, shuffle=True,
        num_workers=n_workers, pin_memory=True, collate_fn=collate,
    )
    val_loader = DataLoader(
        val_ds, batch_size=batch_size, shuffle=False,
        num_workers=n_workers, pin_memory=True, collate_fn=collate,
    )
    test_loader = DataLoader(
        test_ds, batch_size=batch_size, shuffle=False,
        num_workers=n_workers, pin_memory=True, collate_fn=collate,
    )
    
    print(f"[Data] {dataset_name}: train={train_size}, val={val_size}, test={len(test_ds)}")
    return train_loader, val_loader, test_loader


def _get_sensor_size(dataset_name: str) -> Tuple[int, int, int]:
    """Return (H, W, polarities) for known DVS datasets."""
    sizes = {
        "nmnist": (34, 34, 2),
        "dvs_gesture": (128, 128, 2),
        "n_caltech101": (240, 180, 2),
    }
    return sizes.get(dataset_name, (34, 34, 2))


def _event_collate_fn(batch):
    """
    Custom collate for event-based data.
    Handles variable-size tensors from tonic datasets.
    """
    data = []
    targets = []
    for item in batch:
        x, y = item
        if isinstance(x, np.ndarray):
            x = torch.from_numpy(x).float()
        elif isinstance(x, torch.Tensor):
            x = x.float()
        data.append(x)
        targets.append(y)
    
    data = torch.stack(data, dim=0)
    targets = torch.tensor(targets, dtype=torch.long)
    return data, targets


def _fallback_mnist(data_dir: str):
    """Fallback: use torchvision MNIST with flat pixel values."""
    from torchvision import datasets, transforms
    
    transform = transforms.Compose([
        transforms.ToTensor(),
        transforms.Lambda(lambda x: x.view(-1)),  # Flatten to (784,)
    ])
    
    train_ds = datasets.MNIST(data_dir, train=True, download=True, transform=transform)
    test_ds = datasets.MNIST(data_dir, train=False, download=True, transform=transform)
    
    return train_ds, test_ds


def preprocess_for_beam_net(x: torch.Tensor, cfg: dict) -> torch.Tensor:
    """
    Preprocess batch for BEAM-Net input.
    
    For event frames: aggregate temporal bins → first-spike latency map.
    For flat images: normalize to [0, 1].
    
    Parameters
    ----------
    x : torch.Tensor
        Raw batch from dataloader.
    cfg : dict
        Experiment config.
    
    Returns
    -------
    processed : torch.Tensor, shape (batch, d_flat)
        Flattened, normalized input ready for the encoder.
    """
    if x.dim() == 4:
        # Event frames: (batch, n_bins, 2, H, W) → flatten spatial
        # Take ON polarity, compute first-spike time per pixel
        on_events = x[:, :, 0, :, :]  # (batch, n_bins, H, W)
        has_event = on_events > 0
        any_event = has_event.any(dim=1)
        first_bin = has_event.float().argmax(dim=1).float()
        n_bins = x.shape[1]
        processed = first_bin / n_bins
        processed[~any_event] = 1.0  # No event → late spike
        processed = processed.reshape(x.shape[0], -1)  # (batch, H*W)
    elif x.dim() == 5:
        # (batch, n_bins, 2, H, W)
        on_events = x[:, :, 0, :, :]
        has_event = on_events > 0
        any_event = has_event.any(dim=1)
        first_bin = has_event.float().argmax(dim=1).float()
        n_bins = x.shape[1]
        processed = first_bin / n_bins
        processed[~any_event] = 1.0
        processed = processed.reshape(x.shape[0], -1)
    elif x.dim() == 2:
        # Already flat: (batch, d)
        processed = x
    else:
        # (batch, C, H, W) → flatten
        processed = x.reshape(x.shape[0], -1)
    
    # Normalize to [0, 1]
    batch_min = processed.min(dim=-1, keepdim=True)[0]
    batch_max = processed.max(dim=-1, keepdim=True)[0]
    processed = (processed - batch_min) / (batch_max - batch_min + 1e-8)
    
    return processed