requirements.txt
# ============================================================================
# BEAM-Net Requirements
# ============================================================================
# NOTE: torch + torchvision are installed separately in the Dockerfile
#       using the CPU-only index to avoid the 2.5 GB GPU build.
# ============================================================================

# Core system (pkg_resources lives here — required by MLflow)
setuptools>=69.0.0

# Core scientific stack
numpy==1.24.4
scipy==1.11.4

# Parquet + columnar analytics (lakehouse pattern)
pyarrow==14.0.2
duckdb==0.9.2

# Spiking neural networks
spikingjelly==0.0.0.0.14
snntorch==0.7.0
tonic==1.4.3

# Experiment tracking & orchestration
mlflow==2.11.3
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
# Pin flask-session — Airflow 2.7.x breaks with v0.6.0+ due to module rename
flask-session==0.5.0

# Storage
psycopg2-binary==2.9.9
boto3==1.34.0
minio==7.2.0

# Visualization & reporting
matplotlib==3.8.2
seaborn==0.13.0

# Report generation
fpdf2==2.7.6

# Config & utilities
pyyaml==6.0.1
pandas==2.1.4
scikit-learn==1.3.2
tqdm==4.66.1